nlp-module with translation
I can translate STOPWORDS if need. because translation it's slow, is better implement to have a dictionary that will be updated if it's a new language examinated. the dictionary able me to translate only once the language. 
Translate every time a text, because is too long, process will be too slow